+++ include/weather.hpp
@@ -150,7 +150,7 @@ private:
 class Weather : public QObject {
     Q_OBJECT
     Q_PROPERTY(QString city READ get_city NOTIFY city_updated)
-    Q_PROPERTY(float temp READ get_temperature NOTIFY temperature_changed)
+    Q_PROPERTY(double temp READ get_temperature NOTIFY temperature_changed)
     Q_PROPERTY(QUrl weatherIcon READ get_weather_icon_url NOTIFY icon_changed)
 public:
     /**
